update: build_firmware.py support read method with input arguments update: wifi tx optimization to send batch content rather than a few words.

+++ grove_drivers/grove_example/grove_example_class.h
@@ -21,7 +21,7 @@ public:
     bool read_with_arg(float *cx, float *cy, float *cz, int *degree, int arg);
     bool write_acc_mode(uint8_t mode);
     bool write_float_value(float f);
-    bool write_multi_value(int a, float b, int8_t c);
+    bool write_multi_value(int a, float b, uint32_t c);
     bool attach_event_handler(CALLBACK_T handler);
 
 private:
